Update ENTROPIC_SEER_X.7
An ambitious attempt at a reasoning engine. the problem is it is so compressed and so dense that i don't have the informational reserves to disable the native filters to make it work well grrrrr. advances will be made and i will deploy the full version but for now this is pretty gimp